Refactor skill tooling and split out swebench skill
