+++ crates/alan/src/cli/mod.rs
@@ -5,6 +5,7 @@ pub mod daemon;
 pub mod init;
 pub mod shell;
 pub mod skill_authoring;
+pub mod skill_eval;
 pub mod skills;
 pub mod workspace;
