+++ crates/runtime/skills/repo-coding/references/package.md
@@ -9,7 +9,7 @@ is defined in `docs/spec/alan_coding_steward_contract.md`.
 ## Package Principles
 
 This package should make Alan better at general repo-local coding work. It
-should not evolve into a SWE-bench adapter with product-specific behavior.
+must not absorb benchmark orchestration or corpus-specific behavior.
 
 Rules:
 
@@ -35,77 +35,34 @@ Rules:
 5. `references/delivery_contract.md` and `references/evaluator_boundary.md`
    describing the bounded output contract and conditional evaluator boundary.
 6. `evals/evals.json`, `evals/files/benchmark_cases.json`, and
-   `evals/files/delivery_contract_*.json` for manifest-first benchmark and
+   `evals/files/delivery_contract_*.json` for manifest-first local eval and
    delivery-contract scaffolding.
-7. `scripts/` with deterministic validators and benchmark helpers.
+7. `scripts/` with deterministic validators and local benchmark helpers.
 8. External smoke and harness entrypoints under `scripts/repo-worker/` and
    `scripts/harness/`.
 
-## Quick validation
+## Quick Validation
 
 1. `bash scripts/repo-worker/run_smoke.sh`
 2. `bash scripts/harness/run_repo_worker_suite.sh --ci-blocking`
 3. `bash scripts/harness/run_coding_steward_suite.sh --ci-blocking`
 4. `cargo run -p alan -- skills eval crates/runtime/skills/repo-coding --output-dir target/skills-eval/repo-coding/latest`
 
-## External Benchmark Bring-Up
+## External Benchmarks
 
-The Lite-first full-steward external benchmark path starts with a single
-operator-run case:
+SWE-bench and similar operator-run external benchmark adapters no longer live
+in this package. They are owned by the separate `swebench` first-party skill:
 
-1. prepare one clean `SWE-bench Lite` workspace checkout,
-2. prepare the benchmark problem statement text,
-3. fill in `evals/files/swebench_lite_case.template.json`,
-4. run:
+1. `crates/runtime/skills/swebench/bin/swebench-lite-prepare-workspaces`
+2. `crates/runtime/skills/swebench/bin/swebench-lite-materialize-subset`
+3. `crates/runtime/skills/swebench/scripts/run_swebench_full_steward_case.sh`
+4. `crates/runtime/skills/swebench/scripts/run_swebench_full_steward_subset.sh`
+5. `crates/runtime/skills/swebench/scripts/score_swebench_predictions.sh`
 
-```bash
-bash crates/runtime/skills/repo-coding/scripts/run_swebench_full_steward_case.sh \
-  crates/runtime/skills/repo-coding/evals/files/swebench_lite_case.template.json
-```
+That split is intentional:
 
-That runner should:
-
-1. use the steward runtime as the benchmark entrypoint,
-2. require repo-local work to happen through delegated child launch,
-3. export `model.patch` plus single-case prediction artifacts,
-4. emit Alan-native orchestration metadata alongside the benchmark patch.
-
-Those artifacts measure transfer quality for Alan's coding line. They should
-never become the source of benchmark-corpus-specific runtime rules.
-
-The next bring-up step is a curated subset aggregator:
-
-```bash
-python3 crates/runtime/skills/repo-coding/scripts/prepare_swebench_lite_workspaces.py \
-  --instance-ids-file crates/runtime/skills/repo-coding/evals/files/swebench_lite_pilot_v1.ids.txt \
-  --dataset-name princeton-nlp/SWE-bench_Lite \
-  --workspace-root target/benchmarks/swebench_lite/workspaces/pilot_v1
-
-python3 crates/runtime/skills/repo-coding/scripts/prepare_swebench_lite_subset.py \
-  --instance-ids-file crates/runtime/skills/repo-coding/evals/files/swebench_lite_pilot_v1.ids.txt \
-  --dataset-name princeton-nlp/SWE-bench_Lite \
-  --workspace-root target/benchmarks/swebench_lite/workspaces/pilot_v1 \
-  --output-dir target/benchmarks/swebench_lite/manifests/pilot_v1
-
-bash crates/runtime/skills/repo-coding/scripts/run_swebench_full_steward_subset.sh \
-  target/benchmarks/swebench_lite/manifests/pilot_v1/suite.json
-```
-
-That suite runner should:
-
-1. execute each Lite case through the same steward entrypoint,
-2. aggregate suite-level `predictions.jsonl`,
-3. emit `run.json`, `benchmark.json`, `kpi.json`, and `case_results.jsonl`,
-4. surface orchestration counters such as `spawn_count` and `escalation_count`
-   inside the run artifacts,
-5. generate `score_with_official_harness.sh` by delegating to the package-local
-   `score_swebench_predictions.sh` wrapper.
-
-The workspace preparation script intentionally materializes only clean git
-checkouts at each case's `base_commit`. Official resolved/unresolved scoring
-still happens through the Docker-backed SWE-bench harness, while host-native
-dependency provisioning remains operator-owned.
-
-At the owned-output level, reruns are recoverable: stale or partial workspace
-directories are recreated automatically, while `--reuse-existing-workspaces`
-keeps already-clean matching checkouts in place.
+1. `repo-coding` remains the measured coding product.
+2. `swebench` owns benchmark orchestration, dataset adaptation, and official
+   harness scoring.
+3. Full-steward benchmark runs still exercise the real product path:
+   steward session -> `$repo-coding` -> repo-worker child.
