+++ crates/runtime/skills/skill-creator/SKILL.md
@@ -6,7 +6,8 @@ description: |
   Use this when:
   - The user wants to create a new skill package
   - The user wants to update or refactor an existing skill package
-  - You need to structure SKILL.md, sidecars, scripts, references, or assets
+  - You need to structure SKILL.md, sidecars, package-local binaries, scripts,
+    references, or assets
   - You need to validate package shape, execution metadata, or description fit
   - You need to plan explicit authoring and evaluation workflows for a skill
 
@@ -31,8 +32,10 @@ Package-local surfaces:
 
 - `SKILL.md`: portable selection contract and core workflow
 - `skill.yaml` / `package.yaml`: Alan-native runtime defaults
-- `scripts/`: package-private deterministic helpers when a Rust CLI/bin is not
-  the better fit
+- `bin/`: package-local executable tools that should travel with the skill in
+  source and packaged form
+- `scripts/`: package-private glue, wrappers, and compatibility helpers when a
+  `bin/` executable or shared Rust CLI/bin is not the better fit
 - `references/`: material to load only when needed
 - `assets/`: templates or output resources
 - `agents/`: child-agent roots and authoring metadata such as `openai.yaml`
@@ -47,8 +50,9 @@ shape and using explicit tooling.
 2. Pick a short package name in lowercase hyphen-case.
 3. Scaffold the package with `alan skills init`.
 4. Keep `SKILL.md` lean. Move detailed reference material into `references/`.
-5. Prefer reusable Rust CLI/bin tooling first; keep only package-private or
-   ecosystem-bound helpers in `scripts/`.
+5. Prefer package-local `bin/` executables for deterministic skill-private
+   tools, reusable Rust CLI/bin tooling for shared workflows, and keep only
+   glue or ecosystem-bound helpers in `scripts/`.
 6. If the skill delegates, export a package-local child agent under `agents/`.
 7. Validate with `alan skills validate --path <package>`.
 8. Run explicit evaluation with `alan skills eval --path <package>`.
@@ -61,20 +65,20 @@ shape and using explicit tooling.
 - Reuse `assets/templates/` when you need starter package content.
 - Use `alan skills init`, `alan skills validate`, and `alan skills eval` for
   the primary authoring flow.
-- Use the package-local compatibility wrappers under `scripts/` for review
-  artifact regeneration.
+- Use `bin/aggregate-benchmark` and `bin/generate-review` for package-local
+  review artifact regeneration.
 
 ## Rules
 
 1. Prefer one package over sprawling nested abstractions.
-2. Keep runtime tools, package-local helpers, and shared authoring tooling
-   separate.
-3. Prefer Rust CLI/bin surfaces over shell, Python, or TypeScript helpers
-   unless an external ecosystem or a tiny package-private step makes a script
-   the better fit.
+2. Keep host/runtime tools, package-local `bin/` tools, package-local helpers,
+   and shared authoring tooling separate.
+3. Prefer package-local `bin/` executables or shared Rust CLI/bin surfaces over
+   shell, Python, or TypeScript helpers unless an external ecosystem or a tiny
+   package-private step makes a script the better fit.
 4. If shared authoring or eval helpers move into Rust, prefer consolidating
-   them into existing packages such as `alan-tools` or the `alan` CLI rather
-   than introducing another standalone helper package.
+   them into existing operator surfaces such as the `alan` CLI rather than
+   introducing another standalone helper package.
 5. Do not auto-load authoring or eval assets into the runtime prompt.
 6. Make `description` concrete enough that catalog-based selection stays
    reliable.
