+++ crates/runtime/skills/skill-creator/scripts/aggregate_benchmark.py
@@ -1,28 +1,5 @@
-#!/usr/bin/env python3
+#!/usr/bin/env bash
+set -euo pipefail
 
-import argparse
-import shutil
-import subprocess
-import sys
-
-
-def resolve_command() -> list[str]:
-    binary = shutil.which("alan-skill-tools")
-    if binary:
-        return [binary]
-    return ["cargo", "run", "-p", "alan-skill-tools", "--"]
-
-
-def main() -> int:
-    parser = argparse.ArgumentParser(
-        description="Regenerate benchmark.json for an existing eval run directory."
-    )
-    parser.add_argument("run_dir", help="Eval run directory containing run.json")
-    args = parser.parse_args()
-
-    command = [*resolve_command(), "aggregate-benchmark", args.run_dir]
-    return subprocess.run(command, check=False).returncode
-
-
-if __name__ == "__main__":
-    sys.exit(main())
+script_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
+exec "$script_dir/../bin/aggregate-benchmark" "$@"
