+++ crates/runtime/skills/swebench/evals/files/swebench_lite_case.template.json
