+++ crates/runtime/skills/swebench/evals/files/swebench_lite_subset.template.json
