+++ crates/runtime/skills/swebench/references/package.md
@@ -1,0 +1,33 @@
+# SWE-bench Package
+
+This first-party package productizes Alan's SWE-bench benchmark operator layer
+under `crates/runtime/skills/swebench/`.
+
+It is separate from the coding product contract. The measured coding path
+remains Alan steward orchestration plus `$repo-coding` repo-worker delegation.
+
+## What this package contains
+
+1. `SKILL.md` as the parent-facing benchmark operator entry.
+2. `bin/` entrypoints for deterministic workspace preparation and subset
+   materialization.
+3. `scripts/` for full-steward case/subset execution and official harness
+   scoring.
+4. `evals/files/` templates and curated instance-id lists for Lite-first runs.
+5. `tooling/` as the colocated workspace crate that builds the SWE-bench helper
+   binaries used by this package.
+
+## Quick entrypoints
+
+```bash
+crates/runtime/skills/swebench/bin/swebench-lite-prepare-workspaces ...
+crates/runtime/skills/swebench/bin/swebench-lite-materialize-subset ...
+bash crates/runtime/skills/swebench/scripts/run_swebench_full_steward_case.sh ...
+bash crates/runtime/skills/swebench/scripts/run_swebench_full_steward_subset.sh ...
+bash crates/runtime/skills/swebench/scripts/score_swebench_predictions.sh ...
+```
+
+## Boundary
+
+This package is the benchmark operator layer. It should not become the place
+where repo-local coding behavior is defined or specialized.
