+++ crates/runtime/skills/swebench/tooling/src/lib.rs
@@ -1,0 +1,9 @@
+mod swebench_lite;
+
+pub use swebench_lite::{
+    MaterializeSwebenchLiteSubsetOptions, MaterializeSwebenchLiteSubsetResult,
+    PrepareSwebenchLiteWorkspacesOptions, PrepareSwebenchLiteWorkspacesResult,
+    SwebenchLitePreparationEntry, SwebenchLitePreparationFailure,
+    SwebenchLiteSubsetMaterializationReport, SwebenchLiteWorkspacePreparationReport,
+    materialize_swebench_lite_subset, prepare_swebench_lite_workspaces,
+};
