--- crates/skill-tools/src/lib.rs
@@ -1,9 +1,0 @@
-mod eval;
-
-pub use eval::{
-    SkillEvalBenchmarkSummary, SkillEvalCaseManifest, SkillEvalCaseRunSummary,
-    SkillEvalCommandRunSummary, SkillEvalComparisonManifest, SkillEvalComparisonMode,
-    SkillEvalManifest, SkillEvalReviewManifest, SkillEvalRunOptions, SkillEvalRunSummary,
-    SkillEvalStageManifest, default_eval_manifest_path, generate_review_bundle, load_eval_manifest,
-    regenerate_benchmark, run_eval_manifest,
-};
